requirements.txt
flask
opencv-python
torch
torchvision
numpy
Pillow
transformers
